requirements.txt
torch
torchvision
torchaudio
transformers
datasets
accelerate
peft
bitsandbytes
unsloth
wandb
pillow
opencv-python
tqdm
python-dotenv
pyyaml
scikit-learn
evaluate
gradio
matplotlib
graphviz
pypandoc